% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cast_arrow_dtype}
\alias{cast_arrow_dtype}
\title{Cast an Arrow Table variable to Another Data Type}
\usage{
cast_arrow_dtype(arrow_table, var_name, data_type)
}
\arguments{
\item{arrow_table}{An \code{Arrow Table}.}

\item{var_name}{Variable which would be casted to another data type.}

\item{data_type}{An \code{Arrow} data type.}
}
\value{
An \code{Arrow Table}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Change the data type of an Arrow Table column.
}
\examples{
srvdata = readRDS(system.file("extdata", "srvdata.rds", package = "utils.ninsoc"))
srvdata_arrow = arrow::arrow_table(srvdata)
cast_arrow_dtype(srvdata_arrow, Died, arrow::int16())

}
\author{
Fabio M. Vaz
}
