% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compress_arrow}
\alias{compress_arrow}
\title{Compress Arrow Table Variable's Data Types}
\usage{
compress_arrow(x, int64 = FALSE, exclude = NULL)
}
\arguments{
\item{x}{A \code{data.frame} or an \code{Arrow Table}.}

\item{int64}{If \code{TRUE} then big integers could be converted to 64bit integers.
Otherwise it would be converted to double precision (float64).}

\item{exclude}{Vector containing variable's names exclusion list. Those variables will not be
compressed.}
}
\value{
Returns an \code{Arrow Table} with columns converted to
the most memory saving data type.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Convert each \code{Arrow Table} column to the most memory saving data type.
This function doesn't change strings, date or datetime columns.
}
\examples{
srvdata = readRDS(system.file("extdata", "srvdata.rds", package = "utils.ninsoc"))
dplyr::glimpse(arrow::Table$create(srvdata))
dplyr::glimpse(compress_arrow(srvdata))

}
\author{
Fabio M. Vaz
}
