% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compress_data}
\alias{compress_data}
\title{Compress Data.Frame Variable's Data Types}
\usage{
compress_data(x)
}
\arguments{
\item{x}{A \code{data.frame} object.}
}
\value{
Returns a compressed \code{data.frame}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Convert each \code{data.frame} column to the most memory saving data type.
This function doesn't change date or datetime columns, but it can convert
a string column into a numeric one.
}
\examples{
srvdata = readRDS(system.file("extdata", "srvdata.rds", package = "utils.ninsoc"))
dplyr::glimpse(srvdata)
dplyr::glimpse(compress_data(srvdata))

}
\author{
Fabio M. Vaz
}
