% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_variables}
\alias{df_variables}
\title{Get Variable Names and Types from a Data Frame}
\usage{
df_variables(x)
}
\arguments{
\item{x}{A \code{data.frame}.}
}
\value{
Returns a \code{data.frame} containing the following columns:
\describe{
\item{var}{Variable name}
\item{type}{Variable type (\code{integer}, \code{double}, \code{Date}, \code{character}, \code{logical}, \code{factor}, \code{datetime})}
}

Additional attributes include:
\itemize{
\item \code{"num_rows"}: Number of rows in \code{x}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Extracts variable names and their simplified data types from a data frame.
}
\examples{
srvdata = readRDS(system.file("extdata", "srvdata.rds", package = "utils.ninsoc"))
df_variables(srvdata)

}
\seealso{
\code{\link{fst_variables}}
}
\author{
Fabio M. Vaz
}
