% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fst_variables}
\alias{fst_variables}
\title{Variables Names and Types of a \code{fst} File}
\usage{
fst_variables(file)
}
\arguments{
\item{file}{Path to a \code{fst} file.}
}
\value{
Returns a \link[tibble:tibble-package]{tibble} containing the following columns:
\describe{
\item{var}{Column name}
\item{type}{Column type (\code{integer}, \code{double}, \code{Date}, \code{character}, \code{logical}, \code{factor}, \code{datetime})}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Return the name and type of each \code{fst} file columns.
}
\examples{
fst_variables(system.file("extdata", "srvdata.fst", package = "utils.ninsoc"))

}
\seealso{
\code{\link{df_variables}}, \code{\link{pq_variables}}
}
\author{
Fabio M. Vaz
}
