% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label_values}
\alias{label_values}
\title{Label Values}
\usage{
label_values(x, format, na = "NA")
}
\arguments{
\item{x}{Vector to be transformed.}

\item{format}{Named vector structured as c('value' = 'description').}

\item{na}{String to be interpreted as missing values.}
}
\value{
Factor variable.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Create a factor variable using a named vector as a value-description pair.
}
\details{
The 'factor' function in R is weird because it asks you to inform both factor
levels and labels as separated vectors. This can be a problem for factors with
many levels, as the chance of errors arrising from different vector dimensions
or misaligment between labels and values is righer. The function 'label_values'
try to resolve this by asking the user to inform instead a value-description
format that can be made by specifying a named vector.
}
\examples{
srvdata = readRDS(system.file("extdata", "srvdata.rds", package = "utils.ninsoc"))
table(srvdata$Died)
fmt_death = c("TRUE" = "Dead", "FALSE" = "Alive")
srvdata$Died = label_values(srvdata$Died, fmt_death)
table(srvdata$Died)

}
\seealso{
\code{\link[base]{factor}}
}
\author{
Fabio M. Vaz
}
