% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optimal_chunk_size}
\alias{optimal_chunk_size}
\title{Define an Optimal Chunk Size for a Parquet file}
\usage{
optimal_chunk_size(x, chunk_size_bytes = 500 * (1024^2))
}
\arguments{
\item{x}{A \code{data.frame} or an \code{Arrow Table}.}

\item{chunk_size_bytes}{Expected memory size (in bytes) for each chunk. Default 500Mb.}
}
\value{
Returns the optimal number of rows for each chunk.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Define an optimal chunk size (in number of rows) given a max memory partition.
This can be used, for example, to define the chunk size of a parquet file.
Chunk sizes can be defined using the formula \code{chunk_size_bytes = size*(1024^unit)},
where 'unit' can be 1 = Kb, 2 = Mb, 3 = Gb and so on.
}
\examples{
srvdata = readRDS(system.file("extdata", "srvdata.rds", package = "utils.ninsoc"))
utils::object.size(srvdata)
nrow(srvdata)
srvdata_size_by_four = as.numeric(utils::object.size(srvdata))/4
optimal_chunk_size(srvdata, chunk_size_bytes = srvdata_size_by_four)

}
\author{
Fabio M. Vaz
}
