% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{parse_sas_input_code}
\alias{parse_sas_input_code}
\title{Create a Data Dictionary from a SAS Input Code}
\source{
This function is an adaptation of the parse.SAScii function from package SAScii
made by Anthony Joseph Damico (https://cran.r-project.org/package=SAScii).
}
\usage{
parse_sas_input_code(
  sas_input_file,
  beginline = 1,
  lrecl = NULL,
  encoding = NULL
)
}
\arguments{
\item{sas_input_file}{Path to a SAS imput code file.}

\item{beginline}{Line number in SAS import instructions where the INPUT statement begins.  If the word INPUT appears before the actual INPUT block, the function will return an error.}

\item{lrecl}{LRECL option from SAS code.  Only necessary if the width of the ASCII file is longer than the actual columns containing data (if the file contains empty space on the right side).}

\item{encoding}{File encoding (defaults to "cp1252"). See example.}
}
\value{
This function returns a \code{data.frame} including columns:
\describe{
\item{varname}{The name of the variable field}
\item{width}{The width of the field}
\item{char}{A logical flag indicating a character field if TRUE and numeric if FALSE}
\item{divisor}{A fraction to later be multiplied by numeric fields containing decimal points}
\item{start}{Field start position}
\item{end}{Field ending position}
}
}
\description{
Create a \code{data.frame} with columns name, type and widths of a SAS file
retrieved from a SAS input code.
}
\details{
This function cannot handle overlapping columns.
Essa função funciona para a maioria, mas não para todas as situações.
Por exemplo, a função falha quando a leitura de um arquivo (INPUT) é interrompida
por um "@;" para posteriormente ser retomada por outro comando INPUT. Nesse caso
o codigo de importação SAS tem que ser editado previamente para que a função funcione.
}
\seealso{
\code{\link{sas_input_dict}}
}
\author{
Fabio M. Vaz
}
