% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibge_pnadc.R
\name{pnadc_design_lowcase}
\alias{pnadc_design_lowcase}
\title{Cria um objeto de survey.design da PNADC Trimestral ou Anual}
\usage{
pnadc_design_lowcase(data_pnadc)
}
\arguments{
\item{data_pnadc}{A tibble of PNADC microdata read with \code{read_pnadc} function.}
}
\value{
An object of class \code{survey.design} or \code{svyrep.design} with the data from PNADC and its sample design.
}
\description{
This function creates PNADC survey object with its sample design for analysis using \code{survey} package functions.
}
\note{
Essa função foi adaptada da função 'pnadc_design' do pacote "PNADcIBGE".
}
\author{
Fabio M. Vaz
}
