% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{sas_input_dict}
\alias{sas_input_dict}
\title{List of Data Dictionaries from a SAS Input Code when more than one Input is present in the same
Input Code.}
\usage{
sas_input_dict(sas_input_file, file_ext = "txt", encoding = NULL)
}
\arguments{
\item{sas_input_file}{Caminho completo para o script de leitura SAS.}

\item{file_ext}{Extensão dos arquivos de dados (normalmente "txt" ou "dat").}

\item{encoding}{Codificação do arquivo.}
}
\value{
This function returns a list whose entry is the filename and whose
value is a \code{data.frame} with the following columns:
\describe{
\item{varname}{The name of the variable field}
\item{width}{The width of the field}
\item{char}{A logical flag indicating a character field if TRUE and numeric if FALSE}
\item{divisor}{A fraction to later be multiplied by numeric fields containing decimal points}
\item{start}{Field start position}
\item{end}{Field ending position}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This script iterate over a SAS input code, identify each input file, parse
the input code for each file and return a list contaning a data dictionary
for each file.
}
\seealso{
\code{\link{parse_sas_input_code}}
}
\author{
Fabio M. Vaz
}
