% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.ninsoc-package.R
\docType{package}
\name{utils.ninsoc-package}
\alias{utils.ninsoc}
\alias{utils.ninsoc-package}
\title{utils.ninsoc: Utils Package from Ninsoc/Ipea}
\description{
Utils package from Ninsoc/Ipea for general use in data analysis.
}
\author{
\strong{Maintainer}: Fabio Vaz \email{fabio.vaz@ipea.gov.br} (\href{https://orcid.org/0000-0003-0580-0747}{ORCID})

Authors:
\itemize{
  \item Fabio Vaz \email{fabio.vaz@ipea.gov.br} (\href{https://orcid.org/0000-0003-0580-0747}{ORCID})
}

}
\keyword{internal}
